Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    double float num;
+    float num;
     scanf("%lf", &num);
     printf("You entered: %lf", num);
     return 0;
